afsd: Correctly detect invalid cache dirs
afsd has quite a bit of logic for searching through an existing
OpenAFS cache dir, and removing extra cache files. Part of this logic
currently looks similar to this:

    if ((vFileNum = GetDDirNumber(currp->d_name, maxDir)) >= 0) {
        /* we found a valid D1234 dir */

    } else if ((vFileNum = GetDDirNumber(currp->d_name, 1 << 30)) >= 0) {
        /* the D1234 dir we found isn't valid, but only because its
         * number was too big */
    }

That is, we check if we found a valid D1234 dir, and whether or not
the number for the the dir was too high. We go into different code
paths for these, because for fully-valid dirs, we process them like
normal. For too-big dirs, we don't want to keep the dir, but we still
look inside it to potentially find valid cache files.

The problem is, GetDDirNumber doesn't actually use its second
argument; it is passed to doGetXFileNumber, where the 'maxNum'
argument is never used. Instead, we always check the interpreted
number against 'cacheFiles'; that's valid for V1234 files, but not for
D-dirs.

So, if doSweepAFSCache encounters a D-dir that is only invalid because
the number for it is too large, it will process it as if it were a
normal valid dir. That means we access cache_dir_list[vFileNum] for
that dir, but the 'cache_dir_list' array is only 'maxDir' members
long, so writing to that part of the array overwrites some other part
of memory.

As a result, if afsd is started with a cache that has D-dirs that are
much larger than the max dirs needed for the current cache, it can
easily abort or segfault as a result of memory corruption. This can
happen if a client's cache size is greatly reduced (or some other
cache parameters are changed), and the client is restarted with the
same cache dir.

Reviewed-on: https://gerrit.openafs.org/12612
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
(cherry picked from commit 9c1804495d404b4d5b325a7b87789dd54c7fd446)

Change-Id: I26c2148d090eacba5086fa7c1ea792e4faf6cbb5
Reviewed-on: https://gerrit.openafs.org/16103
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
Tested-by: BuildBot <buildbot@rampaginggeek.com>