SOLARIS: Make gafs_fid take a struct fid*
Currently, gafs_fid takes a 'struct fid **' argument, and gives it to
afs_fid(). But afs_fid() takes a 'struct fid *', and the Solaris
vop_fid callback also takes a 'struct fid *'. Just change the argument
to 'struct fid *' to get rid of compiler warnings for both of these
mismatches.

Change-Id: Ie937e4b24ef51265a5957666eda496661f3943ad
Reviewed-on: https://gerrit.openafs.org/12697
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: Andrew Deason <adeason@sinenomine.net>